requirements.txt
fastapi[all]
motor
pydantic>=2.0
python-dotenv